# ddtrace/tracer: applyPPROFLabels: allocate max-length labels slice