@evanj 2022-08-16
I just checked back on this change, and it still appears relevant, and that the code around it has not changed.